Work in Progress
Problem: Write a program which reads in a series of integers, and which
computes the average of these numbers. The program should stop reading
integers when it has read the number 99999.
NOTE: the final 99999 should NOT be reflected in the average
you calculate, rather it is used as an execute average computation
command.