Wrote program for swap
This was such a nifty way of doing swap (only works in C i'm told) it
was an extra credit question on the final I just took. Unfortunately my
answer was only half correct as I did not think the end part would
execute. But it does and I thought it was so neat I just had to come
home and program it up!